requirements-dev.txt
black
ruff
mypy
pandas-stubs
autoflake
pyupgrade
pre-commit
pytest
pytest-timeout
pytest-rerunfailures
coverage
